durable object deployed indexer
moved a bunch of stuff into utility functions

this is our eventual deploy target unless there is a reason it does not work. we will make delegate statements work here and this is the first step.

given the nature of this workload, performance really suffers when building this application upon existing managed stateful services like dynamo or managed postgres (they lack large batch updates) and managing a stateful service is not something anyone should chose to do.

next steps:
1. write something to batch load data from r2
2. write something to batch write data to r2
3. move the /graphql endpoint to a durable object
4. hook up background process running every few seconds to update storage
5. delegate statements