deploy stuff
running into issues with the cpu time limit and this massive batch load. it is going to need to be incrementalized.

it is time to rethink approach. no datastore can handle massive batch writes (short of copying a leveldb file). this might be an anti-pattern for a fundamental reason, not something else