Add Relay Lint Rule for Unused Fields (#64)
* add eslint rule for graphql queries

this has the relay/unused-fields lint rule which ensures we're only reading fields we actually use. this is important so we can better understand the query pattern we're working with here instead of guessing

* add eslint-disable around route queries

it seems like code-splitting doesn't work correctly currently. should fix this eventually but will deal with that separately

* handle existing lint errors

wow it caught some garbage unused things nice