fix liquid delegation bugs (#196)
* fix display for max redelegation not 0 or 1 or 255

* ban console

i hate putting this rule in but the fact we have random console.log everywhere shows we're not able to maintain hygiene without it

* cleanup bugs introduced in #99

what an innocently named pr. it has a few problems. in general, we like to avoid putting if statements in leaf nodes and copying them everwhere when it is possible to constrain the state space at a higher level. concretely

1. pluralizeAddress and pluralizeDelegations should remove other elements than just the text in the case zero. the place to handle the none case is outside this function not in it.

2. in DelegatedToListRow we don't want to push if statements into the render part of this function and then special case edge cases everywhere. this is how bugs happen

3. these random divs wrapping things do nothing

if i'm here fixing pre-launch things i need to review every pr.

* always show delegating to

hiding these is really confusing, we need to show this value in all cases value is flowing through the edge but these rules exist regardless of whether value flows through them

these rules affect both redelegation and individual delegation, redelegation is outside of the delegates control

* fix remove delegation buttons missing

they were being skipped because self-delegation was enabled